Finish Day 5
The trick here is that you have to make use of a Range-Set. Basically the given the list out of start and end vectors, you need to figure out what the new start and end vectors are based on the mapping. This means looping through the Range Set and arriving at four cases for which the overlapping but not completely ones need to be split into two separate ranges. Sounds confusing, so here's a simple example:

Suppose we want to convert temp to humidity. We reach a range where we have (20-50) as the start and end for one of the values in the current vector that holds the ranges for temp. When we ask the range-set for humidity, we get that values from (15-25) temp  are mapped to (10-20) humidity, so this means that we know that at-least temp for (20-25) maps to (15-20) humidity. And now, we need to just find the conversion for (26-50) using similar logic.