Finish Day 7
Not too bad, though, I don't enjoy the fact that I had to pass in a boolean for part two. There was defs a better way to clean the code after part one, but passing the boolean was the quickest way to get the answer. An idea would be to be able to provide your own mapping, but seems pretty annoying and complicated to write up tbh.